Move gov keys to types package
